fix compile issues [wip]
